Move header footer related methods to extension
